+++ config/console/bootstrappers.php
@@ -2,6 +2,7 @@
 use Opulence\Framework\Composer\Bootstrappers\ComposerBootstrapper;
 use Opulence\Framework\Console\Bootstrappers\CommandsBootstrapper as OpulenceCommandsBootstrapper;
 use Opulence\Framework\Console\Bootstrappers\RequestBootstrapper;
+use Opulence\Framework\Databases\Bootstrappers\MigrationBootstrapper;
 use Project\Application\Bootstrappers\Console\Commands\CommandsBootstrapper as ProjectCommandsBootstrapper;
 use Project\Application\Bootstrappers\Http\Routing\RouterBootstrapper;
 use Project\Application\Bootstrappers\Http\Views\ViewBootstrapper;
@@ -17,5 +18,6 @@ return [
     RequestBootstrapper::class,
     ComposerBootstrapper::class,
     ViewBootstrapper::class,
-    ProjectCommandsBootstrapper::class
+    ProjectCommandsBootstrapper::class,
+    MigrationBootstrapper::class
 ];