+++ src/Application/Bootstrappers/Databases/SqlBootstrapper.php
@@ -33,7 +33,8 @@ class SqlBootstrapper extends Bootstrapper implements ILazyBootstrapper
     {
         try {
             $connectionPool = new SingleServerConnectionPool(
-                new Driver(), new Server(
+                new Driver(),
+                new Server(
                     getenv('DB_HOST'), getenv('DB_USER'), getenv('DB_PASSWORD'), getenv('DB_NAME'), getenv('DB_PORT')
                 )
             );