+++ src/Infrastructure/Databases/Migrations/CreateExecutedMigrationTable.php
@@ -1,0 +1,9 @@
+<?php
+namespace Project\Infrastructure\Databases\Migrations;
+
+use Opulence\Framework\Databases\Migrations\CreateExecutedMigrationTable as BaseMigration;
+
+class CreateExecutedMigrationTable extends BaseMigration
+{
+    // Don't do anything unless you want to override the built-in migration
+}